move client to api/client
