+++ node/node.go
@@ -18,13 +18,13 @@ import (
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/params"
 	"github.com/ethereum/go-ethereum/rlp"
+	"github.com/harmony-one/harmony/api/client"
+	clientService "github.com/harmony-one/harmony/api/client/service"
 	proto_node "github.com/harmony-one/harmony/api/proto/node"
 	"github.com/harmony-one/harmony/api/services/explorer"
 	"github.com/harmony-one/harmony/api/services/syncing"
 	"github.com/harmony-one/harmony/api/services/syncing/downloader"
 	downloader_pb "github.com/harmony-one/harmony/api/services/syncing/downloader/proto"
-	"github.com/harmony-one/harmony/client"
-	clientService "github.com/harmony-one/harmony/client/service"
 	bft "github.com/harmony-one/harmony/consensus"
 	"github.com/harmony-one/harmony/core"
 	"github.com/harmony-one/harmony/core/types"
